Extract STARTER_FOLDER constants to a separate module
