+++ src/backend/base/langflow/initial_setup/constants.py
@@ -1,0 +1,2 @@
+STARTER_FOLDER_NAME = "Starter Projects"
+STARTER_FOLDER_DESCRIPTION = "Starter projects to help you get started in Langflow."
